+++ src/lib.rs
@@ -129,26 +129,13 @@ pub use types::{
 /// An identifier in Rhai. [`SmartString`](https://crates.io/crates/smartstring) is used because most
 /// identifiers are ASCII and short, fewer than 23 characters, so they can be stored inline.
 #[cfg(not(feature = "internals"))]
-#[cfg(not(feature = "no_smartstring"))]
 pub(crate) type Identifier = SmartString;
-
-/// An identifier in Rhai.
-#[cfg(not(feature = "internals"))]
-#[cfg(feature = "no_smartstring")]
-pub(crate) type Identifier = ImmutableString;
 
 /// An identifier in Rhai. [`SmartString`](https://crates.io/crates/smartstring) is used because most
 /// identifiers are ASCII and short, fewer than 23 characters, so they can be stored inline.
 #[cfg(feature = "internals")]
-#[cfg(not(feature = "no_smartstring"))]
 #[deprecated = "this type is volatile and may change"]
 pub type Identifier = SmartString;
-
-/// An identifier in Rhai.
-#[cfg(feature = "internals")]
-#[cfg(feature = "no_smartstring")]
-#[deprecated = "this type is volatile and may change"]
-pub type Identifier = ImmutableString;
 
 /// Alias to [`Rc`][std::rc::Rc] or [`Arc`][std::sync::Arc] depending on the `sync` feature flag.
 pub use func::Shared;
@@ -314,11 +301,7 @@ type StaticVec<T> = smallvec::SmallVec<[T; 3]>;
 #[cfg(feature = "internals")]
 pub type StaticVec<T> = smallvec::SmallVec<[T; 3]>;
 
-#[cfg(not(feature = "no_smartstring"))]
 pub(crate) type SmartString = smartstring::SmartString<smartstring::LazyCompact>;
-
-#[cfg(feature = "no_smartstring")]
-pub(crate) type SmartString = String;
 
 // Compiler guards against mutually-exclusive feature flags